Try to fix tests
- Fix double youtube package
