+++ .gitignore
@@ -1,5 +1,7 @@
 .gradle
 /build/
+.idea
+local.properties
 
 # Ignore Gradle GUI config
 gradle-app.setting
