+++ src/test/java/org/schabi/newpipe/extractor/services/youtube/YoutubeSearchEngineChannelTest.java
@@ -1,10 +1,4 @@
-package org.schabi.newpipe.extractor.services.youtube.youtube;
-
-import static junit.framework.Assert.assertEquals;
-import static junit.framework.Assert.assertFalse;
-import static junit.framework.Assert.assertTrue;
-
-import java.util.EnumSet;
+package org.schabi.newpipe.extractor.services.youtube;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -13,6 +7,12 @@ import org.schabi.newpipe.extractor.InfoItem;
 import org.schabi.newpipe.extractor.NewPipe;
 import org.schabi.newpipe.extractor.search.SearchEngine;
 import org.schabi.newpipe.extractor.search.SearchResult;
+
+import java.util.EnumSet;
+
+import static junit.framework.Assert.assertEquals;
+import static junit.framework.Assert.assertFalse;
+import static junit.framework.Assert.assertTrue;
 
 
 /**
@@ -38,7 +38,7 @@ import org.schabi.newpipe.extractor.search.SearchResult;
 /**
  * Test for {@link SearchEngine}
  */
-public class YoutubeSearchEngineStreamTest {
+public class YoutubeSearchEngineChannelTest {
     private SearchResult result;
 
     @Before
@@ -46,8 +46,10 @@ public class YoutubeSearchEngineStreamTest {
         NewPipe.init(Downloader.getInstance());
         SearchEngine engine = NewPipe.getService("Youtube").getSearchEngineInstance();
 
-        result = engine.search("this is something boring", 0, "de",
-                EnumSet.of(SearchEngine.Filter.STREAM)).getSearchResult();
+        // Youtube will suggest "gronkh" instead of "grrunkh"
+        // keep in mind that the suggestions can change by country (the parameter "de")
+        result = engine.search("grrunkh", 0, "de",
+                EnumSet.of(SearchEngine.Filter.CHANNEL)).getSearchResult();
     }
 
     @Test
@@ -57,11 +59,12 @@ public class YoutubeSearchEngineStreamTest {
 
     @Test
     public void testChannelItemType() {
-        assertEquals(result.resultList.get(0).info_type, InfoItem.InfoType.STREAM);
+        assertEquals(result.resultList.get(0).info_type, InfoItem.InfoType.CHANNEL);
     }
 
     @Test
     public void testResultErrors() {
+        if (!result.errors.isEmpty()) for (Throwable error : result.errors) error.printStackTrace();
         assertTrue(result.errors == null || result.errors.isEmpty());
     }
