+++ src/test/java/org/schabi/newpipe/extractor/services/youtube/YoutubeSearchEngineStreamTest.java
@@ -1,16 +1,18 @@
-package org.schabi.newpipe.extractor.services.youtube.youtube;
-
-import static junit.framework.Assert.assertFalse;
-import static junit.framework.Assert.assertTrue;
-
-import java.util.EnumSet;
+package org.schabi.newpipe.extractor.services.youtube;
 
 import org.junit.Before;
 import org.junit.Test;
 import org.schabi.newpipe.Downloader;
+import org.schabi.newpipe.extractor.InfoItem;
 import org.schabi.newpipe.extractor.NewPipe;
 import org.schabi.newpipe.extractor.search.SearchEngine;
 import org.schabi.newpipe.extractor.search.SearchResult;
+
+import java.util.EnumSet;
+
+import static junit.framework.Assert.assertEquals;
+import static junit.framework.Assert.assertFalse;
+import static junit.framework.Assert.assertTrue;
 
 
 /**
@@ -36,7 +38,7 @@ import org.schabi.newpipe.extractor.search.SearchResult;
 /**
  * Test for {@link SearchEngine}
  */
-public class YoutubeSearchEngineAllTest {
+public class YoutubeSearchEngineStreamTest {
     private SearchResult result;
 
     @Before
@@ -44,9 +46,10 @@ public class YoutubeSearchEngineAllTest {
         NewPipe.init(Downloader.getInstance());
         SearchEngine engine = NewPipe.getService("Youtube").getSearchEngineInstance();
 
-        result = engine.search("asdf", 0, "de",
-                EnumSet.of(SearchEngine.Filter.CHANNEL,
-                        SearchEngine.Filter.STREAM)).getSearchResult();
+        // Youtube will suggest "results" instead of "rsults",
+        // keep in mind that the suggestions can change by country (the parameter "de")
+        result = engine.search("rsults", 0, "de",
+                EnumSet.of(SearchEngine.Filter.STREAM)).getSearchResult();
     }
 
     @Test
@@ -55,7 +58,13 @@ public class YoutubeSearchEngineAllTest {
     }
 
     @Test
+    public void testChannelItemType() {
+        assertEquals(result.resultList.get(0).info_type, InfoItem.InfoType.STREAM);
+    }
+
+    @Test
     public void testResultErrors() {
+        if (!result.errors.isEmpty()) for (Throwable error : result.errors) error.printStackTrace();
         assertTrue(result.errors == null || result.errors.isEmpty());
     }
