+++ src/test/java/org/schabi/newpipe/extractor/services/youtube/YoutubeStreamExtractorDefaultTest.java
@@ -1,4 +1,4 @@
-package org.schabi.newpipe.extractor.services.youtube.youtube;
+package org.schabi.newpipe.extractor.services.youtube;
 
 import static junit.framework.Assert.assertTrue;
